+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 1. Added support to Connection Templates Resource.
    - GET /rest/connection-templates/{id}
    - PUT /rest/connection-templates/{id}
+2. Added support to Appliance SSH Access Resource.
+   - GET  /rest/appliance/ssh-access
+   - PUT /rest/appliance/ssh-access
 
 ### Notes
 - This release supports API3000 minimally where we can use OneView v6.20 with this SDK.